+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 Hydrophone is the module responsible for sending emails.
 This API sends notifications to users for things like forgotten passwords, initial signup, and invitations.
 
+## Unreleased
+### Added
+- PT-899: Preview emails
+
 ## 0.7.0 - 2019-10-18
 ### Added
 - PT-671: Display the application version number on the status endpoint (/status).